pyproject.toml
[tool.poetry]
name = "infinitearithmetic"
version = "0.1.0"
description = ""
authors = ["Justin <jbtcao@gmail.com>"]

[tool.poetry.dependencies]
python = "*"
lark-parser = "^0.6.5"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
